orchestration_src/requirements.txt
functions-framework==3.3.0
requests==2.28.2
google==3.0.0
google-api-python-client==2.86.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
prefect==2.10.5
google-cloud-storage==2.8.0
PyYAML==6.0
